GridTable Now Accepts Params and can be used by anyone as a shared component
