+++ src/Components/Analysis/GridTables/GridTable.js
@@ -1,57 +1,13 @@
-import { Avatar } from '@mui/material';
 import Box from '@mui/material/Box';
 import { DataGrid, GridToolbar } from '@mui/x-data-grid';
-import LatestOrders from './LatestOrders.json';
-const columns = [
-  { field: 'id', headerName: 'OrderID', width: 90 ,type: 'number'},
-  {
-    field: 'user',
-    headerName: 'User Name',
-    width: 150,
-    flex:2,
-    renderCell: (params) => {
-      return (
-        <>
-          <Avatar src={params.value.avatar}/>
-            {params.value.username}
-        </>
-      );
-    }
-  },
-  {
-    field: 'item-name',
-    headerName: 'Item Name',
-    width: 110,
-    flex:2
-  },
-  {
-    field: 'quantity',
-    headerName: 'Quantity',
-    width: 110,
-    flex:1,
-  },
-  {
-    field: 'order-total',
-    headerName: 'Total',
-    width: 110,
-    flex:1
-  },
-  {
-    field: 'date',
-    headerName: 'Date',
-    width: 150,
-    type:'date',
-  },
-];
-
-const GridTable =() => {
+const GridTable =({tableTitle,cols,data}) => {
   return (
     <Box sx={{ height: 400, width: '100%' }}>
-      Latest Orders
+      {tableTitle}
       <DataGrid
-        rows={LatestOrders}
+        rows={data}
         disableDensitySelector
-        columns={columns}
+        columns={cols}
         components={{ Toolbar: GridToolbar }}
         componentsProps={{
           toolbar: {
